requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1

# Security and authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI and Image Processing
torch==2.1.0
torchvision==0.16.0
numpy==1.26.1
opencv-python==4.8.1.78
pillow==10.1.0
scikit-image==0.22.0
matplotlib==3.8.1

# XAI
grad-cam==1.4.8
shap==0.43.0

# Report generation
reportlab==4.0.7
fpdf==1.7.2
jinja2==3.1.2

# Testing
pytest==7.4.3
httpx==0.25.1

# CORS
starlette==0.27.0

# Validation and parsing
email-validator==2.1.0.post1
